+++ packages/core/src/render3/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {createComponentRef, detectChanges, getHostElement, getRenderedText, markDirty, renderComponent, whenRendered} from './component';
+import {createComponentRef, getHostElement, getRenderedText, markDirty, renderComponent, whenRendered} from './component';
 import {NgOnChangesFeature, PublicFeature, defineComponent, defineDirective, definePipe} from './definition';
 import {InjectFlags} from './di';
 import {ComponentDef, ComponentTemplate, ComponentType, DirectiveDef, DirectiveDefFlags, DirectiveType} from './interfaces/definition';
@@ -64,6 +64,7 @@ export {
 
   embeddedViewStart as V,
   embeddedViewEnd as v,
+  detectChanges,
 } from './instructions';
 
 export {
@@ -109,7 +110,6 @@ export {
   defineComponent,
   defineDirective,
   definePipe,
-  detectChanges,
   createComponentRef,
   getHostElement,
   getRenderedText,
